+++ packages/cibul-node/api/middleware/redirectIfPrivate.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = (req, res, next) => {
+  if (req.agenda.private) {
+    return res.redirect(302, `/v2/agendas/${req.agenda.uid}.prv`);
+  }
+  next();
+}